examples/v2/multi-trigger/requirements.txt
azure-functions
azure-functions-doctor
